Add support for pruned models (#103)
* Add support for pruned models

According to the [pytorch documentation on
pruning](https://pytorch.org/tutorials/intermediate/pruning_tutorial.html),
the original parameter is replaced with one ending with `_orig` and
a new buffer ending with `_mask`. The mask contains 0s and 1s based on
which the correct parameters are chosen.

All instances of `param.nelements()` have been replaced by a variable
`cur_params` whose value is set based on whether it is a masked model or
not. To keep consistency with the rest of the code base, the `_orig` is
removed from the `name` variable right after the calculation of
`cur_params`.

* Add tests for pruning